Moved the configuration to a folder
