+++ BL/getGroups.js
@@ -1,5 +1,5 @@
 const axios = require('axios');
-const config = require('./generateConfig');
+const config = require('../config/generateConfig');
 
 const GROUPS_URL = global.gConfig.getGroupsURL
